add merchant_client with mock test

+++ lib/features/merchant/data/remote/models/coordinates.dart
@@ -6,8 +6,8 @@ part 'coordinates.g.dart';
 @freezed
 class Coordinates with _$Coordinates {
   factory Coordinates({
-    @JsonKey(name: 'latitude') int? latitude,
-    @JsonKey(name: 'longitude') int? longitude,
+    @JsonKey(name: 'latitude') double? latitude,
+    @JsonKey(name: 'longitude') double? longitude,
   }) = _Coordinates;
 
   factory Coordinates.fromJson(Map<String, dynamic> json) =>
